+++ src/aux-common/bots/test/BotCalculationContextTests.ts
@@ -1122,23 +1122,23 @@ export function botCalculationContextTests(
     describe('getPortalConfigBotID()', () => {
         it('should return the bot ID that the config bot tag points to', () => {
             const userBot = createBot('userBot', {
-                auxPagePortal: 'abc',
+                pagePortal: 'abc',
                 auxPagePortalConfigBot: 'test',
             });
 
             const calc = createPrecalculatedContext([userBot]);
-            const id = getPortalConfigBotID(calc, userBot, 'auxPagePortal');
+            const id = getPortalConfigBotID(calc, userBot, 'pagePortal');
 
             expect(id).toEqual('test');
         });
 
         it('should return null if the tag does not exist', () => {
             const userBot = createBot('userBot', {
-                auxPagePortal: 'abc',
+                pagePortal: 'abc',
             });
 
             const calc = createPrecalculatedContext([userBot]);
-            const id = getPortalConfigBotID(calc, userBot, 'auxPagePortal');
+            const id = getPortalConfigBotID(calc, userBot, 'pagePortal');
 
             expect(id).toEqual(null);
         });